requirements.txt
discord.py==2.4.0
openai==0.27.8
python-dotenv==1.0.0
PlexAPI==4.15.0